+++ oi_cli2/cli/config.py
@@ -1,12 +1,12 @@
 #!/usr/bin/env python3
 import click
 from oi_cli2.core.DI import DI_DB, DI_LOGGER
-from oi_cli2.cli.constant import OT_FOLDER, OT_LOG, USER_CONFIG_FILE
+from oi_cli2.cli.constant import OT_FOLDER, USER_CONFIG_FILE
 from oi_cli2.cli.template import template
 from oi_cli2.cli.account import account
-from oi_cli2.utils.Logger import getLogger
 from oi_cli2.utils.configFolder import ConfigFolder
 from oi_cli2.utils.db import JsonFileDB
+import oi_cli2.core.provider as provider
 import logging
 
 
@@ -17,7 +17,7 @@ def config(ctx):
   # TODO make it provider?
   config_folder = ConfigFolder(OT_FOLDER)
   try:
-    logger: logging = getLogger(config_folder.get_file_path(OT_LOG))
+    logger: logging.Logger = provider.o.get(DI_LOGGER)
   except Exception as e:
     print(str(e))
     exit(1)